+++ src/printer.rs
@@ -219,11 +219,10 @@ impl<'a> InteractivePrinter<'a> {
     pub(crate) fn new(
         config: &'a Config,
         assets: &'a HighlightingAssets,
+        theme: &'a Theme,
         input: &mut OpenedInput,
         #[cfg(feature = "git")] line_changes: &'a Option<LineChanges>,
     ) -> Result<Self> {
-        let theme = assets.get_theme(&config.theme);
-
         let background_color_highlight = theme.settings.line_highlight;
 
         let colors = if config.colored_output {